Extracting `#[endpoint]` code from `function.rs` to `endpoint.rs`
